+++ topobank/files/models.py
@@ -3,6 +3,7 @@ Basic models for handling files and folders, including upload/download logic.
 """
 
 import logging
+from datetime import datetime
 
 from django.conf import settings
 from django.core.exceptions import SuspiciousFileOperation
@@ -59,6 +60,7 @@ class Folder(PermissionMixin, models.Model):
         if not created:
             manifest.file.delete()
         manifest.file = fobj
+        manifest.upload_confirmed = datetime.now()
         manifest.save()
 
     def exists(self, filename: str) -> bool:
@@ -83,6 +85,20 @@ class Folder(PermissionMixin, models.Model):
     def remove_files(self):
         """Clear this folder by removing all files"""
         self.files.all().delete()
+
+    def deepcopy(self, permissions=None):
+        copy = Folder.objects.get(pk=self.pk)
+        copy.pk = None
+        if permissions is not None:
+            copy.permissions = permissions
+        copy.save()
+
+        for filemanifest in self.get_valid_files():
+            filemanifest_copy = filemanifest.deepcopy(permissions=permissions)
+            filemanifest_copy.folder = copy
+            filemanifest_copy.save()
+
+        return copy
 
 
 # The Flow for "direct file upload" is heavily inspired from here:
@@ -294,7 +310,8 @@ class Manifest(PermissionMixin, models.Model):
             # it does not support absolute paths if we use this method.
             try:
                 storage_path = default_storage._normalize_name(
-                    self.generate_storage_path())
+                    self.generate_storage_path()
+                )
             except SuspiciousFileOperation:
                 # This happens after migrations, when the file name is not yet set
                 _log.info(